refactor: 자동차 클래스 디렉토리 이동
