+++ src/main/java/racingcar/domain/car/CarEngine.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.car;
 
 public interface CarEngine {
     boolean move();
